Updating to fully embrace JSON and removing legacy code from character Also turning character object into a roller object
